Main menu was debugged
