+++ init_widgets.cpp
@@ -5,15 +5,17 @@
 #include "rect_widget.hpp"
 #include "text_widget.hpp"
 #include "environment.hpp"
+#include "menu_widget.hpp"
+#include "widget.hpp"
 
 #include <iostream>
 
-void test_event() {
+void test_event(Widget* null) {
     std::cout << "Test output\n";
 }
 
 void init_widgets() {
     widgets.push_back(new ImageWidget(sprites[0], point{400, 0}));
     widgets.push_back(new AnimatedClose(point{1216, 0}));
-    widgets.push_back(new ButtonWidget(fonts[0], 12, sf::Color::Cyan, "test button", test_event, sf::Color::White, { 325,313 }, { 12,12 }));
+    widgets.push_back(new MenuWidget());
 }
