+++ menu_widget.hpp
@@ -7,17 +7,12 @@ class MenuWidget : public Widget {
 	MM_MainPage* main_page;
 	MM_SettingsPage* settings_page;
 
-	MenuWidget();
 public:
+	MenuWidget();
+	void on_click();
 	void main_to_settings();
 	void settings_to_main();
 };
 
-void menu_event_mtos(Widget* menu) {
-	auto menu_ = (MenuWidget*)menu;
-	menu_->main_to_settings();
-}
-void menu_event_stom(Widget* menu) {
-	auto menu_ = (MenuWidget*)menu;
-	menu_->settings_to_main();
-}
+void menu_event_mtos(Widget* menu);
+void menu_event_stom(Widget* menu);
